requirements.txt
asgiref==3.5.2
attr==0.3.1
blinker==1.4
brotli==1.0.9
brotlicffi==1.0.9.2
ConfigParser==5.2.0
cryptography==37.0.2
Cython==0.29.30
dl==0.1.0
docutils==0.18.1
HTMLParser==0.0.2
ipython==8.4.0
ipywidgets==7.7.0
jnius==1.1.0
keyring==23.6.0
lockfile==0.12.2
lxml==4.9.0
mock==4.0.3
numpy==1.22.4
ordereddict==1.1
Pillow==9.1.1
protobuf==4.21.1
pyOpenSSL==22.0.0
python-dotenv==0.20.0
railroad==0.5.0
simplejson==3.17.6
Sphinx==5.0.2
toml==0.10.2
tornado==6.1
trove_classifiers==2022.5.19
unicodedata2==14.0.0
watchdog==2.1.9
xmlrpclib==1.0.1
